+++ src/main/java/co/invest72/investment/presentation/response/CalculateInvestmentResponse.java
@@ -1,5 +1,6 @@
 package co.invest72.investment.presentation.response;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 import lombok.Builder;
@@ -9,7 +10,7 @@ import lombok.Getter;
 @EqualsAndHashCode
 @Getter
 public class CalculateInvestmentResponse {
-	private final int totalInvestment;
+	private final BigDecimal totalInvestment;
 	private final int totalInterest;
 	private final int totalTax;
 	private final int totalProfit;
@@ -19,7 +20,7 @@ public class CalculateInvestmentResponse {
 	private final List<YearlyInvestmentResult> yearlyDetails;
 
 	@Builder
-	public CalculateInvestmentResponse(int totalInvestment, int totalInterest, int totalTax,
+	public CalculateInvestmentResponse(BigDecimal totalInvestment, int totalInterest, int totalTax,
 		int totalProfit, String taxType, String taxPercent, List<MonthlyInvestmentResult> monthlyDetails,
 		List<YearlyInvestmentResult> yearlyDetails) {
 		this.totalInvestment = totalInvestment;
